Ghosts cross portals when chasing Pacman.
Added a few new conditions in the 'move_smartly' function in
'Ghost.java' file. These ensure that when any one of the ghosts are
chasing Pacman (i.e. the 'state' is 'smart') and are crossing a portal
to the other side of the map, they cross is correctly as opposed to
changing to the opposite direction when they see Pacman at the other end
of the map after he crossed first.