Modify bonus score presentation code.
Changed the code regarding bonus score presentation within the game
(i.e. numbers that appear and flash green and white when pacman eats a
ghost). Modified the bonus score variables from boolean type to int and
added a new variable names 'bonusScore' that takes the value of each one
of the previously mentioned variables. Therefore, the bonus scores no
longer function with flags (removing the danger of two of these flags
being active at the same time).

Also replaced the 'make_xxx_appear' functions with one single function
that receives the current bonus score as an argument and is assigned to
the new variable 'bonusScore'.