Create new 'GhostEaten' function in 'Pacman.java' class.
In order to replace the previous four functions that checked if each
individual ghost had been eaten by pacman or not, a general 'GhostEaten'
function was created in order to eliminate the repeated code in the four
previously mentioned functions.

Also renamed a few variables in this class in order to follow comply
with camel case. Trying to keep all code modular.