Create 'isGhostEaten' function in 'Pacman.java' class.
Created this function in order to assess if a ghost has been eaten or
not by pacman when he is energized. This enabled the removal of several
other functions which accomplished the same feat but were more
inefficient. This function still needs a lot of work though. Will
probably be the subject of the next commit(s).