Create new 'HowToPlayPanel' and 'Files' folder.
Created new 'HowToPlayPanel' GUI Panel with the future purpose of
containing information to be displayed to the user on how to play the
pacman game as well as information regarding the game characters,
levels, energizers, portals and more.

Also restructured the 'res' folder. As well as the 'Images' folders it
now also contains a new subfolder named 'Files' where the files
regarding usernames and scores ('Usernames.txt' and 'Scores.txt') have
been moved.