Reverted to 3rd last commit (previous two were broken).
Reverted to this commit (a7fd4bf8cd0347e6d79b7e988a87928107c7b293) as
the previous two showed a few mistakes. Probably were not tested to the
full when committing.

In this commit revert back and solely added the previously mentioned
'ghostArray' containing all ghost objects. This has been tested and does
not seem to be the cause of the previous problems.