Added new 'eatGhost' function.
This function manages the events regarding a ghost being eaten by
pacman. This functions allows for the removal of the previous
'checkEatenGhosts' and 'isGhostEaten' functions which weren't very
explicit in what they did. Ghosts all spawn in the same spot for now
which is not the end goal. Will sort this issue out in a future commit.