Add siren sound effect when playing game with sound.
Added 'loop' function in 'Sounds.java' class in order to make this work.
Just like any other sound effect, in order for the siren sound to be
audible the user must enable the game sound in the settings. This sound
effect should play continuously when the game is running which is
already implemented. However, it keeps playing when the win or lose
screen is presented which should not happen. When these screens are
presented the siren sound effect should stop and this still needs to be
fixed.