Create new 'Zone' class in 'Ghost.java' class.
This new class has been created within the 'Ghost.java' class with the
intent of changing the 'ZonesArray' array to 'Zone' as opposed to
boolean. This was done once that a zone has more information than can
fit into a boolean variable. Ghost AI still seems to be working fine
after this change.