Make a draft for the 'moveSmartly' function in 'Ghost.java' class.
This function now works pretty well along with the 'moveRandomly'
function and the ghosts are able to switch between movement types fairly
smoothly. They still get stuck in specific situation but this will be
addressed in further commits. The goal is to have a smooth AI for the
ghosts similar to the previous existing one but where the code is
significantly more readable.