Fix ghost movement when finding path to Pacman.
Previously, ghosts would get stuck and be immobilised when pacman was
either directly above, below, to the left or to the right of a ghost and
the first direction taken when finding the path was blocked. This has
now been fixed and the ghosts do not seem to get stuck anymore.

On another note, when following pacman through a portal, ghosts now turn
back after pacman spawns on the opposite side of the map. This had
previously been fixed but it seems to be an issue again. Next commit
will probably be to fix this.