Death animation in the correct place.
Pacman's death animation happens in the correct place. However, ghosts
are still present upon death and pacman loses a life each time the
animation ends (and restarts after and so on...) due to the ghosts
responsible for the initial collision still being there and intersecting
pacman. Still need to remove the ghosts upon death, get animation to run
once and then restart the game after (or end if it was pacman's last
life).