Death animation now only runs once after pacman death.
Previously animation would run continuously. And, even though changing
the game status to 'play' after death, sometimes there would be enough
time for the animation to restart before removing it and moving on to
the gameplay. A new 'deathAnimationDisplayed' flag was created in order
to prevent this from happening and now the animation only runs once
without restarting before gameplay is resumed.