Ghosts remain in spawn box for a determined time before exiting.
This was added seen as if Pacman ate a Ghost close to the spawn box door
and the ghost immediately left the spawn box, pacman would die
instantly. This did not make much sense, therefore a delay is now
present (3 seconds) which the ghosts much serve before leaving the spawn
box. However, this delay is only applied after a ghost is eaten and not
at the very beginning of a level run as pacman starts sufficiently far
away from the ghosts to warrant an initial safety period.