Improved ghost movement when not allowed to leave spawn box.
Previously, ghost's movement in the spawn box would be a little "clunky"
after having been eaten by pacman and forced to stay in the box for the
determined period of time before being allowed to exit. This was because
the 'generateNextDirection' function generated any of the 4 possible
directions of travel (including upwards). Therefore the ghost would try
to travel upwards but the 'canLeaveBox' flag would not yet be set to
true, forcing the ghost to stay in the box looking upwards and
motionless. Now, up until 'canLeaveBox' is true the ghosts simply travel
vertically until being allowed out.