New argument 'movementStatus' added to the 'Ghost' class.
This new argument was added seen as it was noted that often ghosts would
not continue to move methodically (chase pacman) after having crossed a
portal. This is because after spawning upon crossing the portal their
movement type was reset back to the 'randomMovement' movement type in
the 'Ghost' constructor. Therefore, 'movementType' is now set to the
'movementStatus' argument which is passed to the constructor before a
ghost is respawn via a portal or after having been eaten by pacman.