Made 'Food' class abstract.
Turned this class abstract since it should only be accessed by its
subclasses. This enabled the removal of the render function from this
class since it was only needed in its subclasses. Also made new abstract
method in this class called 'getFoodScore' which is implemented in the
subclasses (Pellet and Energizer) and solely returns the score value of
each one of these objects.