Pacman memorizes next direction of travel upon crossing portal.
Previously pacman did not memorize next direction of travel to be taken
when this was set before crossing a portal. Meaning that after crossing
the portal pacman would simply continue travelling left or right
(depending from what side he was crossing the portal from), forcing the
user to input the next direction to be taken a further time. Now this
has been resolved and the user can input the next direction he wishes
pacman to travel in even before he has crossed the portal.