Reverted to last working commit.
Noticed that I missed a few bugs in the commits I did after adding more
functionalities. So had to get rid of them and go back to the last
working version. Bugs were that ghosts would not go up due to spawn box
logic and portal crossing of ghosts made vulnerable when already in the
portal was faulty.