'Character' class cleaning continued...
Still a few things I'd like to improve but cannot think of how at the
moment. Duplicated code regarding movement being one of them. More
specifically in the 'canMove' and 'move' functions, which have similar
implementation, although one is checking if the next character position
is valid and the other is actually changing this position after knowing
it is valid. So not sure if its actually duplicated or not. Also, after
cleaning the next classes ('Pacman' and 'Ghost') the movement functions
in 'Character' may suffer more changes which could reduce the mentioned
duplicated code.