Cleaned up 'DeadPacman' class by adding new function.
The new function called 'checkGameStatusNeedsChanging' prevents the need
to change the state of the game in the animation timing function of this
class. It also required the need to create a new flag in order to know
when the death animation of pacman had finished in order to assert when
the game status should change back to 'lose' (if the number of lives is
not 0) or 'play' otherwise.