Ghosts can cross portal from right side only for now.
Need to change the logic behind portal crossing functions. But have been
able to reduce the number of constructors for HostileGhost class by
allowing the x and y spawn coordinates to be passed from the location in
the code where the new hostile ghost must be spawned. Because only there
do they know where to actually spawn the ghost. No other portal
variables will be required to spawn ghosts.